add PokeBattleTestCase helper
